requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
decorator==4.4.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
pycparser==2.20
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
tornado==6.0.4
urllib3==1.25.10
Werkzeug==1.0.1
